✨ Add WichtelMemberController methods

+++ app/Http/Controllers/WichtelMemberController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Group;
 use App\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Hash;
 
 class WichtelMemberController extends Controller
 {
@@ -31,7 +32,41 @@ class WichtelMemberController extends Controller
      */
     public function store(Request $request, Group $group)
     {
-        //
+        // @TODO: Return JSON on validation fail. Should happen automatically
+        $this->validate($request, [
+            'name' => 'required|max:255',
+            'email' => 'required|email',
+            'wishlist' => 'string|max:1000',
+        ]);
+
+        // Check if user is already in DB
+        $member = User::where('email', $request->email)->first();
+
+        if ($member->belongsToGroup($group)) {
+            return response()->json([
+                'message' => 'User already belongs to group.'
+            ]);
+        }
+
+        // If not create new instance
+        if (is_null($member)) {
+            $member = User::create([
+                'name' => $request->name,
+                'email' => $request->email,
+                'password' => Hash::make(generateRandomString(16)),
+            ]);
+        }
+
+        $pivotData = [
+            'wishlist' => $request->wishlist,
+            'status' => 'invited',
+        ];
+
+        $group->users()->save($member, $pivotData);
+
+        // @TODO: dispatch( Invite Mail )
+
+        return response()->json($member, 201);
     }
 
     /**
@@ -43,7 +78,7 @@ class WichtelMemberController extends Controller
      */
     public function show(Group $group, User $wichtelmember)
     {
-        if (! $wichtelmember->belongsToGroup($group)) {
+        if (!$wichtelmember->belongsToGroup($group)) {
             return response()->json([
                 'message' => "This member ({$wichtelmember->id}) is not part of this group ({$group->id})",
             ], 403);
@@ -56,12 +91,27 @@ class WichtelMemberController extends Controller
      * Update the specified resource in storage.
      *
      * @param  \Illuminate\Http\Request $request
-     * @param  int $id
+     * @param Group $group
+     * @param User $wichtelmember
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, $id)
+    public function update(Request $request, Group $group, User $wichtelmember)
     {
-        //
+        if (!$wichtelmember->belongsToGroup($group)) {
+            return response()->json([
+                'message' => "This member ({$wichtelmember->id}) is not part of this group ({$group->id})",
+            ], 403);
+        }
+
+        // @TODO: Return JSON on validation fail. Should happen automatically
+        $this->validate($request, [
+            'wishlist' => 'string|max:1000',
+        ]);
+
+        $wichtelmember->saveWishlist($group, $request->wishlist);
+
+        return response()->json($wichtelmember);
+
     }
 
     /**
@@ -73,13 +123,17 @@ class WichtelMemberController extends Controller
      */
     public function destroy(Group $group, User $wichtelmember)
     {
-        if (! $wichtelmember->belongsToGroup($group)) {
+        if (!$wichtelmember->belongsToGroup($group)) {
             return response()->json([
                 'message' => "This member ({$wichtelmember->id}) is not part of this group ({$group->id})",
             ], 403);
         }
 
-        $wichtelmember->delete();
+        $group->users()->detach($wichtelmember);
+
+        if (count($wichtelmember->groups) === 0) {
+            $wichtelmember->delete();
+        }
 
         return response('', 204);
     }
